IAnGogh/pyproject.toml
[project]
name = "iangogh"
version = "0.1.0"
description = "MVP local employability master system with UiPath integration contracts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pymongo>=4.8.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.2",
  "jinja2>=3.1.4",
  "cryptography>=43.0.0",
  "mongomock>=4.1.2"
]

[tool.uvicorn]
factory = false
host = "127.0.0.1"
port = 8787
